ver1.0

+++ internal/models/response.go
@@ -2,6 +2,6 @@ package models
 
 type FetchResponse struct {
 	Ticker     string  `json:"ticker"`
-	Price      float32 `json:"price"`
-	Difference float32 `json:"difference"`
+	Price      float64 `json:"price"`
+	Difference float64 `json:"difference"`
 }
